refactor: extract out picker module
